change from float32 to float64

+++ _handledata.py
@@ -30,7 +30,7 @@ class DataHandler:
     def get_loaded(self) -> str:
         return self.model
 
-    def get_axis(self, ax: str, vdsrc: str, gateL: str) -> npt.NDArray[np.float32]:
+    def get_axis(self, ax: str, vdsrc: str, gateL: str) -> npt.NDArray[np.float64]:
         match ax:
             case "gmro":
                 return self.__get_gmro(vdsrc, gateL)
@@ -41,21 +41,21 @@ class DataHandler:
             case _:
                 return self.__get_simple(ax, vdsrc, gateL)
     
-    def __get_simple(self, ax: str, vdsrc: str, gateL: str) -> npt.NDArray[np.float32]:
+    def __get_simple(self, ax: str, vdsrc: str, gateL: str) -> npt.NDArray[np.float64]:
         # This regex string is a steaming pile of shit, but it essentially ANDs 4 conditions
         regex_str: str = "(?=.*M0:{})(?=.*vds={})(?=.*length={})(?=.*Y)".format(ax, vdsrc, gateL).replace("+", "\\+")
         return self.df.filter(regex=regex_str).to_numpy()
     
-    def __get_gmro(self, vdsrc: str, gateL: str) -> npt.NDArray[np.float32]:
-        gm:  npt.NDArray[np.float32] = self.__get_simple("gm ", vdsrc, gateL)
-        gds: npt.NDArray[np.float32] = self.__get_simple("gds", vdsrc, gateL)
+    def __get_gmro(self, vdsrc: str, gateL: str) -> npt.NDArray[np.float64]:
+        gm:  npt.NDArray[np.float64] = self.__get_simple("gm ", vdsrc, gateL)
+        gds: npt.NDArray[np.float64] = self.__get_simple("gds", vdsrc, gateL)
         return gm / gds
     
-    def __get_idw(self, vdsrc: str, gateL: str) -> npt.NDArray[np.float32]:
-        id: npt.NDArray[np.float32] = self.__get_simple("id ", vdsrc, gateL)
+    def __get_idw(self, vdsrc: str, gateL: str) -> npt.NDArray[np.float64]:
+        id: npt.NDArray[np.float64] = self.__get_simple("id ", vdsrc, gateL)
         return id / 1e-6
     
-    def __get_ft(self, vdsrc: str, gateL: str) -> npt.NDArray[np.float32]:
-        gm:  npt.NDArray[np.float32] = self.__get_simple("gm ", vdsrc, gateL)
-        cgg: npt.NDArray[np.float32] = self.__get_simple("cgg ", vdsrc, gateL)
+    def __get_ft(self, vdsrc: str, gateL: str) -> npt.NDArray[np.float64]:
+        gm:  npt.NDArray[np.float64] = self.__get_simple("gm ", vdsrc, gateL)
+        cgg: npt.NDArray[np.float64] = self.__get_simple("cgg ", vdsrc, gateL)
         return gm / (2 * np.pi * cgg)
